requirements.txt
bcrypt==4.0.1
cryptography==43.0.0
fastapi==0.112.2
psycopg2-binary==2.9.9
python-jose==3.3.0
python-multipart==0.0.9
SQLAlchemy==2.0.32
uvicorn==0.30.6
passlib==1.7.4
httpx==0.27.2
aiofiles==24.1.0
jinja2==3.1.4
python-dotenv==1.0.1
markdown==3.7
bs4==0.0.2
alembic==1.13.2